Remove "at least 2D" rank expansion in fit/predict/evaluate.
Previously, if Model.fit/evaluate/predict saw a batch of data of shape (batch_size,), it would uprank it to become (batch_size, 1). After this change, no upranking is applied.

This is potentially a breaking change for subclassed models that process scalar data.

Note that this does not affect Sequential models defined with an explicit input shape, nor Functional models, because such models will standardize their inputs to what they expect. For instance a model built with `Input(shape=())` that receives data of shape (batch_size, 1) will reshape it to (batch_size,) and inversely a model built with `Input(shape=(1,))` that receives data of shape (batch_size,) will reshape it to (batch_size, 1).

PiperOrigin-RevId: 383542144